feature: integrates context into the handleRequest method of AuthGuard
This feature integrates the context into the handleRequest method of the AuthGuard, so that not only the current user can be checked, but also the user's rights with regard to the context, e.g. with regard to decorators for access rights to resolvers. In addition, a bugfix is integrated which fixes an incorrect parameter sequence in the canActivate method.