feature(@nestjs/passport): add support for basic-auth with passport-http
Currently, when using the BasicAuthStrategy from passport-http, the
parameter "info" is ignored when the callback is called. This leads to an
UnauthorizedException being thrown without the addition of the required
"WWW-Authenticate"-Header in the following response. This behavior causes
the browsers Authentication-Popup to not being shown.
With this commit, the parameter "info" is checked for the string "Basic"
which will be passed in by "passport-http" in order to be added to the
headers of the response. When the authentication was unsuccessful and the
info parameter is set the response will be send including the correct
header, thus enabling the popup window.