fix(core): Set callback function length
Some strategies such as Azure AD use the length of the callback function to
determine the parameters but use of the spread results in a function length
of 0. See https://github.com/AzureAD/passport-azure-ad#5113-verify-callback.
This change sets the length of the callback function from the validate
method plus the done function.